+++ src/scope/repositories/sources.js
@@ -83,12 +83,6 @@ export default class SourceRepository {
     return foundComponent;
   }
 
-  async hasSymlink(bitId: BitId) {
-    const component = Component.fromBitId(bitId);
-    const foundComponent = await this.findComponent(component);
-    return foundComponent instanceof Symlink;
-  }
-
   getObjects(id: BitId): Promise<ComponentObjects> {
     return this.get(id).then((component) => {
       if (!component) throw new ComponentNotFound(id.toString());
@@ -239,7 +233,7 @@ export default class SourceRepository {
       specsResults
     });
     component.addVersion(version, releaseType, exactVersion);
-
+    component.local = true;
     objectRepo.add(version).add(component);
 
     if (files) files.forEach(file => objectRepo.add(file.file));